larbs:emoji: Add rupee sign
This file is the input for the famous Luke Smith script to insert emojis/Unicode characters into anywhere, really.
The symbol I require the most out of any other, however, is the Rupee symbol. I used to insert it using the iso4755 patch in st. 
This commit makes adding the Rupee symbol application-agnostic, as that script works on any application, not just st.