.config:nvim: moveaway from "potato" optimisations
These config options were set when I was using a ThinkPad T60, and as
such are designed to make nvim usable on that potato of a laptop.
Also, it was 32-bit, hence the keybinds to change architecture in AUR
PKGBUILDs.

Now, I'm on a more powerful PC temporarily, and the device I will
hopefully be using in the future will also be much powerful than that
T60, hence eliminating the need to cripple neovim.

Also, having your editor be transparent is *not* a great idea. I wonder
why I thought that would be a good idea.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>