.config:pam-gnupg: Added gerrit ssh keygrip
I was concerned slightly about this file, and security, however,
exposing the keygrip isn't *that* important, like, cruegge himself did in
in the pam-gnupg repo. And only I know the passphrases to these keys
anyway, so no problem.

This adds the keygrip to the ssh key I generated for the TWRP gerrit
instance, so that it gets automatically unlocked as well, through pam-gnupg.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>