.local:bin: T60 trackpoint modifications
Recently, the trackpoint on the T60 stopped working. I have no idea what
caused it, it just did. After troubleshooting for about two weeks, I got
it back working again after unplugging the trackpad from the laptop. The
trackpad was also dead. Maybe that had something to do with it. I don't
know. As a result of the trackpad not existing anymore, the paths to the
files that configure the trackpoint changed, necessitating a change in
the script.

Also, doing "sudo sh -c foo" is considered better than just "sudo foo"
for non-interactive scripts. ¯\_(ツ)_/¯

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>