Replace startx with sx
sx is a more minimalist replacement for startx. For more info, watch
Brodie Robertson's video on the application. TL;DR of the video is:
startx has a bunch of options and configuration that most people,
(including me) never actually use. sx only starts X, nothing else, which
is just what I need. It's a 29-line shell script, which is more secure,
easier to maintain and more minimalist, also, it has OOTB compatibility
with XDG basedir.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>
Change-Id: Id419d35bd335f482d33a05faebcee40b938b4d97