Update README.md
General updates to the readme. 

Since the dotfiles are already initialised, no need to have instructions for the first-time initialisation, also, those instructions can be found elsewhere on the web.

Also, nowadays, I store the .git directory somewhere else, to remove clutter in the homedir, so readme has been updated to reflect that.