Xresources: final changes for readability
All of the Xresources colour values are now good looking, and also have
better readability. The normbg and selbg are similar in colour, so as
to not seem jarring, and yet are different enough to be distinguishable.
And the selfg are now in a different colour from the palette entirely,
to easily see which is the active window/selection.

All of the files have a vim filetype definition at the end, for easier
plugin detection, i.e. vimCSSColor now works on all of the files not
just `Xresources`.

Gruvbox also has a cursor colour definition now.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>
Change-Id: I6addac5f453546903e03bb061736c7b9823cd684