Move to thunderbird
Thunderbird is a much better experience for me now than
neomutt+isync+msmtp.
It handles Oauth2 for gmail, no need for app passwords.
It has ability to send emails in plain text, but also show HTML emails,
which the morons of the world somehow insist on sending.
It also handles my GPG key natively, and so does everything I need it to
do.

It also has calendar and adressbook support natively, which is good,
so no need for calcurse and abook.
(Abook hasn't been updated since 2015. It _is_ feature complete,
but still, security issues. Arch dropped it from their repos
into the AUR recently.)

The only thing I would want the old setup back is for in case I run
across a computer that can't handle Thunderbird. My old T60, for
example.
And also, for sending automated messages to Arch32 about testing
packages, but I don't use Arch32 anymore, so it's mainly a non-issue.
And the T60 broke, so I don't see a chance for me to run 32-bit anytime
soon.

Also, lxpolkit has seemingly been removed from Arch repos, there's a
package in the AUR, but it doesn't compile. So, disable it, and I'm
currently not on a ThinkPad anymore, so disable my script setting the
trackpoint up as well.