gpg: Unset static pinentry program
Due to some advice on the internet I followed, I created my own problem.
Setting a pinentry program in gpg-agent.conf makes it use that
permanently, nad overrides the default.
Because of this, for a long time, I faced the problem for GUI
applications not being able to use GPG, as the password prompt would try
to open in a terminal (curses). And if I set it to a GUI, then it
wouldn't work over ssh.

The default behaviour is that it checks if $DISPLAY is set and
dynamically chooses the appropriate pinentry.

*facepalms* Take internet advice with a grain of salt, people.