zshenv: Move to tty2 as default
On my current system, gdm and (GASP!) GNOME is being used as well as dwm.
The shock and horror! It's not up to me, the system is being shared and
they prefer GNOME. So, I'm starting dwm from tty2 currently, as gdm hogs
tty1.

When I return to a non-shared system, this can be reverted.