sxhkd: Add indicator for volume changes
Before, when I changed the volume by keybindings, I had no idea what the
volume was.
Now, the current volume is displayed in a notification after each
change.
NOTE: The volume display bit is restricted to a single sink, it is very
hacky, as pactl doesn't have a display volume option, so it may or may
not work.