dotfiles: Add wayland setup dotfiles
I recently tried out sway for the first time. It's pretty good, but its
not going to replace dwm anytime soon.
I've got most of my setup working, with tiny things that aren't present
on sway but are on dwm (like the mod+0 workflow, and tiling in general
is different because of the auto vs manual, and the scratchpad)