firefox: Re-enable vaapi decoding for firefox
WebRender has been enabled for Firefox on Linux since FF88, so the
webrender flag is not needed, even if it says false. The RDD sandbox
VAAPI bug has been fixed in 102, so re-enabling VAAPI decoding.

103 will enable it by default, so no need for this flag in the future, I
think, so will remove it in the future if so.

FFVPX is the software decoder for VP8, VP9 inbuilt in Firefox, I'm
guessing, but my GPU has VP8 and VP9 decoding support, so it is not
needed, so disabling it.