.local/bin: Add sway starter script
One of the big problems I had with living the wayland life was that
firefox and thunderbird used to launch using XWayland, despite using
environment variables that forced them to use their native Wayland
renderers.
With this wrapper script, this is now fixed.
My wayland setup is mostly there, with this being fixed, as well as the
autotiling being fixed thanks to the 'autotiling' tool, and the
scratchpad issue has been mostly resolved as well.
I might move to wayland completely soon, idk.