dotfiles: Tentatively move to sway
My sway stuff is mostly done, the only thing I can think of that I'm
missing from dwm, is the swallow functionality. There is a way to
implement it jankily, but there isn't a one-to-one replacement.

The scratchpad is still different from what I'm used to on dwm. dwm
spawns a window when the scratch keybind is pressed, even if I had
ctrl+d'd out of it, sway doesn't. Also, I have to supply multiple
windows myself, not automatic like dwm, and only one keybind which
cycles through them, not like three keybinds for three different
scratchpad windows in dwm.

I still don't have the mod+0 functionality of dwm, but that's fine, I
hardly ever use it now.

Other things are minor, which I'll get used to over time.