waybar: Solarized-ify waybar
The default waybar style.css is renamed to default.css, and the current
config is solarized. The other configs, gruvbox and nord are kanged from
random configs on the interwebs, and I've not gone through them, so
colours might not be matched to the modules. I'll go through them
when/if I switch to that colourscheme.