sway: Add a less janky swallow setup
The hack I was using wasn't proper dwm-like swallow, It used the tabbed
layout to simulate swallow, I wasn't able to tile two windows
side-by-side, and the original terminal window was still visible.

This solution now uses https://github.com/jamesofarell/i3-swallow, which
is the closest we're going to get to a dwm-like swallow natively in sway
or i3. The only breaking changes if I were to switch to something else,
or back to dwm, is that the nnn opener is changed, and isn't compatible
with anything other than sway and i3.

With this, my sway config is as close to my dwm config that I'm going to
get.