sway: Add scratchpad spawn after quitting functionality
The way the scratchpad works in dwm with the scratchpad patches is really cool,
it checks if a window with the specified scratchpad command, or a window in the
scratch tag is present, and if not, it spawns the given scratchpad
command.

So, if I ctrl+d out of the scratchpad terminal, dwm automatically spawns
another the next time I press the scratchpad keybind.

Sway by default doesn't do this, if a window dies in the scratchpad, you
have to move a new one manually, it doesn't do it for you. With this
systemd service, I can automate spawning the scratchpad window, and
automate restarting it after it dies with the `Restart=always`
parameter.

Sway is now practically identical to my dwm workflow, except for the
master stack layout specifically, which I *could* manually make, and the
mod+0 thing, which I don't use. Also, the i3-swallow thing is after all,
a third-party solution, so I can't use mpv's keybinds immediately, I
have to wait for the mpv window to spawn, but there is nothing I can
do about that.

Ditching X permanently is an increasingly viable option, though I'll
still keep a way to easily switch back and forth, indefinitely.