wayland: Add hyprland config
Hyprland is much closer to dwm and my dwm config natively than sway is
after all my configuration.
However, it is currently in beta, in heavy development, and not quite
there yet, I have encountered multiple bugs, instability and some
features are still missing.
When hyprland matures enough, and is stable, and the missing features
are implemented, I'll most probably switch to it. Till then, it's a
project to keep an eye on.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>