hyprland: Add sway-audio-idle-inhibit
Sway has window rules to inhibit idle if a window is fullscreen, but
hyprland doesn't, so swaylock is triggered even when I'm watching
something in Hyprland. This fixes that, and also even on sway, if I'm
listening to something, and the window isn't fullscreen, swaylock is
triggered. This fixes that too.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>