monctrlwlnd: Improve script working
This script now moves from a extremely jank state to a less jank state.
The `keyword monitor disable` command completely removes the monitor
from the monitor stack, and there's no way to bring it back, so
obviously that's not a viable state. So, switched to dpms control
instead, which gives more expected behaviour. But dpms just wakes the
screen up by moving the mouse or by hitting a key, which while being
expected behaviour for dpms itself, is not my usecase's expected
behaviour in this script, the monitor should stay off, until I turn it
on specifically (which is the way it works on X).

So unless hyprland develops a way to disable monitors while not making
them completely unusable and uninteractable after disabling it, this is
the state this script will remain in, unfortunately.

But mirroring works now, so yay for that, I guess. It seems like one
problem has been traded for another, while in fact, the monitor disable
problem probably existed even when I was focused on mirroring not
working, and never tested it this way/was not pointed out in the
hyprland documentation, so really, the mirroring problem was fixed,
while the disabling problem remains from before.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>