hyprland: Add idleinhibit fullscreen rule
This upon basic testing doesn't work as expected, i.e. when mpv video
is paused but fullscreen, and not as according to Sway and dwm behaviour,
IIRC, which did not go into idle in such a state but this might be the
normal expected behaviour. But let's see what happens in the future/in
other test cases.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>