wayland: Drop pasystray for now
Pasystray-wayland and pasystray normal for that matter have not been
updated since a while on Arch, despite 0.8.2 being released a month or
so ago. So switch to pulseaudio module in waybar.

My bluetooth usage has been minimal, so drop blueman applet as well.

Also, fix minor pet peeve that I didn't configure, when clicking mpd
module, it didn't do anything, despite expected behaviour being it to
toggle play-pause. I just had to specify it to do so in the config file.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>