mime: Add changes done by mimeo
Mimeo, one of the many excellent projects by Xyne, helps users easily
manage the fustercluck that is the mime associations and their openers.

This is the result of using it, and as far as I can tell, they are
pretty good now. Further modifications might come as I come across wrong
mime associations.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>