kitty: Add gruvbox and nord themes
These are files grabbed from the kitty +kitten themes menu.
I've modified the background colour in nord and gruvbox
by just a single digit so that it is still very similar to the original
color, but by doing so, it stops making the background transparent in
nvim in kitty.

The solarizedpatched colours are different from the solarized8 theme
anyway, so no theme change is required there.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>