yt-dlp: Use better flag to limit resolution to 1080p
The flag I used to use didn't work all the time, Hardware Unboxed videos
would pull in the 4k stream.
So, use better flag I found, that hopefully fixes this once and for all.
At least the same videos now pull in the correct 1080p stream thanks to
this flag, so this may finally be fixed, let's see.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>